requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
opencv-python==4.9.0.80
ultralytics==8.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.26.3